refactor(privacy): split floor + erasure selectors into own files (500-LOC gate)
The F-011/F-012 additions plus the origin/main merge pushed retention.go
(518) and erasure.go (514) over the repo's hard 500-LOC-per-file cap, failing
the deterministic-gates `go-file-length` check. Rather than waive (the
waivers file is for pre-existing debt, not new growth), split each along a
clean concept boundary — a pure code move within package privacy, no
behaviour change:

- retention_floor.go: correspondenceFloorPredicate, statutoryCorrespondenceFloor,
  statutoryFloorArgs (the shared statutory-floor helpers).
- erasure_selectors.go: subjectOnlyActivities, subjectOnlyDestroyable,
  unlinkedCapturedMail (the person-erase activity-selection SQL).

retention.go → 446, erasure.go → 447. `scripts/check-go-file-length.sh` passes;
go build ./..., go vet ./..., and the privacy unit tests are green.
